django-blog/requirements.txt
certifi==2019.11.28
chardet==3.0.4
coverage==4.5.4
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.40.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==3.0.0
idna==2.8
oauthlib==3.1.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.0
urllib3==1.25.7
